+++ kaggle_sets/datasets/testing/daily-min-temperatures.csv
@@ -1,0 +1,301 @@
+Date,Temp
+0,1981-01-01,20.7
+1,1981-01-02,17.9
+2,1981-01-03,18.8
+3,1981-01-04,14.6
+4,1981-01-05,15.8
+5,1981-01-06,15.8
+6,1981-01-07,15.8
+7,1981-01-08,17.4
+8,1981-01-09,21.8
+9,1981-01-10,20.0
+10,1981-01-11,16.2
+11,1981-01-12,13.3
+12,1981-01-13,16.7
+13,1981-01-14,21.5
+14,1981-01-15,25.0
+15,1981-01-16,20.7
+16,1981-01-17,20.6
+17,1981-01-18,24.8
+18,1981-01-19,17.7
+19,1981-01-20,15.5
+20,1981-01-21,18.2
+21,1981-01-22,12.1
+22,1981-01-23,14.4
+23,1981-01-24,16.0
+24,1981-01-25,16.5
+25,1981-01-26,18.7
+26,1981-01-27,19.4
+27,1981-01-28,17.2
+28,1981-01-29,15.5
+29,1981-01-30,15.1
+30,1981-01-31,15.4
+31,1981-02-01,15.3
+32,1981-02-02,18.8
+33,1981-02-03,21.9
+34,1981-02-04,19.9
+35,1981-02-05,16.6
+36,1981-02-06,16.8
+37,1981-02-07,14.6
+38,1981-02-08,17.1
+39,1981-02-09,25.0
+40,1981-02-10,15.0
+41,1981-02-11,13.7
+42,1981-02-12,13.9
+43,1981-02-13,18.3
+44,1981-02-14,22.0
+45,1981-02-15,22.1
+46,1981-02-16,21.2
+47,1981-02-17,18.4
+48,1981-02-18,16.6
+49,1981-02-19,16.1
+50,1981-02-20,15.7
+51,1981-02-21,16.6
+52,1981-02-22,16.5
+53,1981-02-23,14.4
+54,1981-02-24,14.4
+55,1981-02-25,18.5
+56,1981-02-26,16.9
+57,1981-02-27,17.5
+58,1981-02-28,21.2
+59,1981-03-01,17.8
+60,1981-03-02,18.6
+61,1981-03-03,17.0
+62,1981-03-04,16.0
+63,1981-03-05,13.3
+64,1981-03-06,14.3
+65,1981-03-07,11.4
+66,1981-03-08,16.3
+67,1981-03-09,16.1
+68,1981-03-10,11.8
+69,1981-03-11,12.2
+70,1981-03-12,14.7
+71,1981-03-13,11.8
+72,1981-03-14,11.3
+73,1981-03-15,10.6
+74,1981-03-16,11.7
+75,1981-03-17,14.2
+76,1981-03-18,11.2
+77,1981-03-19,16.9
+78,1981-03-20,16.7
+79,1981-03-21,8.1
+80,1981-03-22,8.0
+81,1981-03-23,8.8
+82,1981-03-24,13.4
+83,1981-03-25,10.9
+84,1981-03-26,13.4
+85,1981-03-27,11.0
+86,1981-03-28,15.0
+87,1981-03-29,15.7
+88,1981-03-30,14.5
+89,1981-03-31,15.8
+90,1981-04-01,16.7
+91,1981-04-02,16.8
+92,1981-04-03,17.5
+93,1981-04-04,17.1
+94,1981-04-05,18.1
+95,1981-04-06,16.6
+96,1981-04-07,10.0
+97,1981-04-08,14.9
+98,1981-04-09,15.9
+99,1981-04-10,13.0
+100,1981-04-11,7.6
+101,1981-04-12,11.5
+102,1981-04-13,13.5
+103,1981-04-14,13.0
+104,1981-04-15,13.3
+105,1981-04-16,12.1
+106,1981-04-17,12.4
+107,1981-04-18,13.2
+108,1981-04-19,13.8
+109,1981-04-20,10.6
+110,1981-04-21,9.0
+111,1981-04-22,10.0
+112,1981-04-23,9.8
+113,1981-04-24,11.5
+114,1981-04-25,8.9
+115,1981-04-26,7.4
+116,1981-04-27,9.9
+117,1981-04-28,9.3
+118,1981-04-29,9.9
+119,1981-04-30,7.4
+120,1981-05-01,8.6
+121,1981-05-02,11.9
+122,1981-05-03,14.0
+123,1981-05-04,8.6
+124,1981-05-05,10.0
+125,1981-05-06,13.5
+126,1981-05-07,12.0
+127,1981-05-08,10.5
+128,1981-05-09,10.7
+129,1981-05-10,8.1
+130,1981-05-11,10.1
+131,1981-05-12,10.6
+132,1981-05-13,5.3
+133,1981-05-14,6.6
+134,1981-05-15,8.5
+135,1981-05-16,11.2
+136,1981-05-17,9.8
+137,1981-05-18,5.9
+138,1981-05-19,3.2
+139,1981-05-20,2.1
+140,1981-05-21,3.4
+141,1981-05-22,5.4
+142,1981-05-23,9.6
+143,1981-05-24,11.5
+144,1981-05-25,12.3
+145,1981-05-26,12.6
+146,1981-05-27,11.0
+147,1981-05-28,11.2
+148,1981-05-29,11.4
+149,1981-05-30,11.8
+150,1981-05-31,12.8
+151,1981-06-01,11.6
+152,1981-06-02,10.6
+153,1981-06-03,9.8
+154,1981-06-04,11.2
+155,1981-06-05,5.7
+156,1981-06-06,7.1
+157,1981-06-07,2.5
+158,1981-06-08,3.5
+159,1981-06-09,4.6
+160,1981-06-10,11.0
+161,1981-06-11,5.7
+162,1981-06-12,7.7
+163,1981-06-13,10.4
+164,1981-06-14,11.4
+165,1981-06-15,9.2
+166,1981-06-16,6.1
+167,1981-06-17,2.7
+168,1981-06-18,4.3
+169,1981-06-19,6.3
+170,1981-06-20,3.8
+171,1981-06-21,4.4
+172,1981-06-22,7.1
+173,1981-06-23,4.8
+174,1981-06-24,5.8
+175,1981-06-25,6.2
+176,1981-06-26,7.3
+177,1981-06-27,9.2
+178,1981-06-28,10.2
+179,1981-06-29,9.5
+180,1981-06-30,9.5
+181,1981-07-01,10.7
+182,1981-07-02,10.0
+183,1981-07-03,6.5
+184,1981-07-04,7.0
+185,1981-07-05,7.4
+186,1981-07-06,8.1
+187,1981-07-07,6.6
+188,1981-07-08,8.3
+189,1981-07-09,8.9
+190,1981-07-10,4.6
+191,1981-07-11,6.8
+192,1981-07-12,5.7
+193,1981-07-13,6.1
+194,1981-07-14,7.0
+195,1981-07-15,7.2
+196,1981-07-16,6.3
+197,1981-07-17,8.8
+198,1981-07-18,5.0
+199,1981-07-19,7.4
+200,1981-07-20,10.1
+201,1981-07-21,12.0
+202,1981-07-22,9.0
+203,1981-07-23,8.9
+204,1981-07-24,9.8
+205,1981-07-25,9.0
+206,1981-07-26,9.2
+207,1981-07-27,7.7
+208,1981-07-28,8.0
+209,1981-07-29,6.1
+210,1981-07-30,3.5
+211,1981-07-31,3.2
+212,1981-08-01,5.7
+213,1981-08-02,7.7
+214,1981-08-03,9.0
+215,1981-08-04,10.0
+216,1981-08-05,6.2
+217,1981-08-06,6.9
+218,1981-08-07,6.5
+219,1981-08-08,6.8
+220,1981-08-09,7.0
+221,1981-08-10,5.2
+222,1981-08-11,3.0
+223,1981-08-12,5.6
+224,1981-08-13,7.9
+225,1981-08-14,9.0
+226,1981-08-15,8.6
+227,1981-08-16,10.3
+228,1981-08-17,10.5
+229,1981-08-18,7.6
+230,1981-08-19,9.7
+231,1981-08-20,12.5
+232,1981-08-21,7.4
+233,1981-08-22,7.9
+234,1981-08-23,3.9
+235,1981-08-24,6.6
+236,1981-08-25,4.6
+237,1981-08-26,7.0
+238,1981-08-27,6.0
+239,1981-08-28,5.5
+240,1981-08-29,8.1
+241,1981-08-30,5.5
+242,1981-08-31,6.2
+243,1981-09-01,8.0
+244,1981-09-02,10.3
+245,1981-09-03,9.8
+246,1981-09-04,9.6
+247,1981-09-05,8.5
+248,1981-09-06,7.5
+249,1981-09-07,11.2
+250,1981-09-08,14.6
+251,1981-09-09,11.7
+252,1981-09-10,7.8
+253,1981-09-11,12.3
+254,1981-09-12,10.1
+255,1981-09-13,11.5
+256,1981-09-14,7.3
+257,1981-09-15,10.9
+258,1981-09-16,14.1
+259,1981-09-17,10.7
+260,1981-09-18,16.9
+261,1981-09-19,10.5
+262,1981-09-20,6.5
+263,1981-09-21,11.0
+264,1981-09-22,6.3
+265,1981-09-23,10.5
+266,1981-09-24,7.2
+267,1981-09-25,7.6
+268,1981-09-26,10.7
+269,1981-09-27,7.8
+270,1981-09-28,9.6
+271,1981-09-29,11.4
+272,1981-09-30,12.4
+273,1981-10-01,8.9
+274,1981-10-02,13.2
+275,1981-10-03,8.6
+276,1981-10-04,6.2
+277,1981-10-05,11.4
+278,1981-10-06,13.2
+279,1981-10-07,14.3
+280,1981-10-08,7.3
+281,1981-10-09,12.9
+282,1981-10-10,7.8
+283,1981-10-11,6.2
+284,1981-10-12,5.6
+285,1981-10-13,10.0
+286,1981-10-14,13.3
+287,1981-10-15,8.3
+288,1981-10-16,10.2
+289,1981-10-17,8.6
+290,1981-10-18,7.3
+291,1981-10-19,10.4
+292,1981-10-20,11.2
+293,1981-10-21,13.2
+294,1981-10-22,11.4
+295,1981-10-23,9.1
+296,1981-10-24,6.6
+297,1981-10-25,8.4
+298,1981-10-26,9.7
+299,1981-10-27,13.2
